탭 클릭 시 애니메이션 구현 (#10)
* feat #8: 탭 클릭 시 애니메이션 구현

- 탭 클릭 시 메뉴 애니메이션 구현
- 현재 카테고리 공통 부모 컴포넌트인 App으로 이동

* chore #8 : 주석 제거

* refactor #8 : 기존 타입을 재활용
