+++ client/src/components/MenuItem.tsx
@@ -1,0 +1,22 @@
+interface MenuProps {
+  menu: Menu;
+}
+
+export default function MenuItem({ menu }: MenuProps) {
+  return (
+    <li>
+      <img src={menu.src} alt={menu.name} />
+      <span>{menu.name}</span>
+      <span>{menu.price}</span>
+      {menu.isPopular && <PopularMark />}
+    </li>
+  );
+}
+
+function PopularMark() {
+  return (
+    <div>
+      <span>인기</span>
+    </div>
+  );
+}
