+++ client/src/components/MenuList.module.css
@@ -1,0 +1,31 @@
+.AnimateIn {
+  animation-name: slide-down-in;
+  animation-duration: 0.5s;
+}
+
+@keyframes slide-down-in {
+  0% {
+    transform: translateY(-2%);
+    opacity: 0;
+  }
+  100% {
+    transform: translateY(0%);
+    opacity: 1;
+  }
+}
+
+.AnimateOut {
+  animation-name: slide-down-out;
+  animation-duration: 0.5s;
+}
+
+@keyframes slide-down-out {
+  0% {
+    transform: translateY(0);
+    opacity: 1;
+  }
+  100% {
+    transform: translateY(2%);
+    opacity: 0;
+  }
+}
